JITrench

JITrench.DiffableFunction

JITrench.GradField

JITrench.Model

JITrench.PNGContainer

JITrench.TrenchObject

JITrench.Variable

Base.getindex

Base.sum

Base.transpose

JITrench._broadcast_to

JITrench.backward!

JITrench.broadcast_to

JITrench.cleargrad!

JITrench.cleargrads!

JITrench.flatten

JITrench.matmul

JITrench.mean_squared_error

JITrench.plot_graph

JITrench.plot_model

JITrench.sum_to

JITrench.DiffableFunction — Type

DiffableFunction

An abstract type that is the parent type of differentiable functions;

all function in JITrench must be children of this type.

Examples

julia> subtypes(DiffableFunction)
24-element Vector{Any}:
 JITrench.Add
 JITrench.BroadcastTo
 JITrench.Broadcasting
 JITrench.Cos
 JITrench.Div
 JITrench.Exp
 JITrench.Flatten
 JITrench.GetIndex
 JITrench.GetIndexGrad
 JITrench.Log
 ⋮
 JITrench.Reshape
 JITrench.Sigmoid
 JITrench.Sin
 JITrench.Sub
 JITrench.Sum
 JITrench.SumTo
 JITrench.Tan
 JITrench.Transpose
 JITrench._Linear

JITrench.GradField — Type

GradField

A structure that contains all the necessary items for automatic differentiation. All function types in JITrench have this type as a field named "grad_field".

JITrench.Model — Type

Model

Abstract type representing a model.

Child types of this abstract type are recognized as models by JITrench.

For more information, refer to the "Neural Networks" section of the documentation (here).

JITrench.PNGContainer — Type

PNGContainer

A structure for handling images. You can display an image by using it as a return value of a cell or by explicitly display(::PNGContainer) in Jupyter.

JITrench.TrenchObject — Type

TrenchObject

An abstract type that is the root of an object as implemented in Specifically, see also subtypes(foo).

julia> subtypes(TrenchObject)
2-element Vector{Any}:
 DiffableFunction
 Variable

JITrench.Variable — Type

Variable(values, [creator, grad, generation, name])

A type that is a variable in JITrench.

For a detailed explanation of this, see the documentation(here).

Only real numbers or arrays consisting of real numbers can be stored.

Examples

julia> Variable(2)
name: nothing 
values: 2
creator: User-Defined(nothing)

julia> Variable([1.2, 3.5])
name: nothing 
values: [1.2, 3.5]
creator: User-Defined(nothing)

Base.getindex — Method

Base.getindex(x::Variable, ind...)

return x.values[ind...] as Variable.

Examples

julia> x = Variable(rand(2, 2))
name: nothing 
values: [0.7050007249265509 0.5075375401538957; 0.9953109600473362 0.8447135817368259]
creator: User-Defined(nothing)

julia> x[1, 2]
name: nothing 
values: 0.5075375401538957
creator: JITrench.GetIndex

julia> x[1, :]
name: nothing 
values: [0.7050007249265509, 0.5075375401538957]
creator: JITrench.GetIndex

Base.sum — Method

sum(x::Variable; dims=nothing, keepdims=false)

Arguments

keepdims

When true, the number of dimensions in the input and output arrays is guaranteed to match.

Examples

julia> x = Variable([1, 2, 3])
name: nothing 
values: [1, 2, 3]
creator: User-Defined(nothing)

julia> y = sum(x)
name: nothing 
values: 6
creator: JITrench.Sum

julia> x = Variable(rand(2, 2, 2))
name: nothing 
values: [0.8036752887616154 0.918293421092931; 0.9092024610524445 0.8788506330169976]

[0.7726260228472543 0.6884237844746439; 0.38214845033860456 0.660455509851009]
creator: User-Defined(nothing)

julia> y = sum(x, keepdims=true)
name: nothing 
values: [6.013675571435501]
creator: JITrench.Sum

julia> y.values
1×1×1 Array{Float64, 3}:
[:, :, 1] =
 6.013675571435501

Base.transpose — Method

transpose(x::Variable)

Examples

julia> x = Variable([1 2; 3 4])
name: nothing 
values: [1 2; 3 4]
creator: User-Defined(nothing)

julia> transpose(x)
name: nothing 
values: [1 3; 2 4]
creator: JITrench.Transpose

JITrench._broadcast_to — Method

_broadcast_to(A, shape)

Apply broadcast to make size(x.values) as shape.

Examples

julia> _broadcast_to([1, 2, 3], (3, 3))
3×3 Matrix{Float64}:
 1.0  1.0  1.0
 2.0  2.0  2.0
 3.0  3.0  3.0

julia> _broadcast_to(1, (2, 2))
2×2 Matrix{Float64}:
 1.0  1.0
 1.0  1.0

JITrench.backward! — Method

backward(y::Variable; retain_grad=false)

Compute the back propagation, y as the end of the computational graph. For a more detailed explanation, see the documentation (here).

julia> x
name: nothing 
values: 1
creator: User-Defined(nothing)

julia> x = Variable(1)
name: nothing 
values: 1
creator: User-Defined(nothing)

julia> y = x +  2
name: nothing 
values: 3
creator: JITrench.Add

julia> backward!(y)

julia> x.grad
name: nothing 
values: 1
creator: User-Defined(nothing)

JITrench.broadcast_to — Method

broadcast_to(x::Variable, shape)

Apply broadcast to make size(x.values) as shape.

Examples

julia> x = Variable([1, 2, 3])
name: nothing 
values: [1, 2, 3]
creator: User-Defined(nothing)

julia> JITrench.broadcast_to(x, (3, 2))
name: nothing 
values: [1.0 1.0; 2.0 2.0; 3.0 3.0]
creator: JITrench.BroadcastTo

JITrench.cleargrad! — Method

cleargrad!(x::Variable)

Reset the Variable's gradient.

julia> x
name: nothing 
values: 1
grad: Variable(1)
creator: User-Defined(nothing)

julia> x.grad
name: nothing 
values: 1
creator: User-Defined(nothing)

julia> cleargrad!(x)

julia> x.grad === nothing
true

JITrench.cleargrads! — Method

cleargrads!(model::Model; skip_uninit=false)

clear information about grad of the layers which is returned from layers. if skip_uninit is true, parameters which is uninit(i.e nothing) is skipped.

JITrench.flatten — Method

flatten(x::Variable)

The function corresponding to vcat(x...).

Example

julia> x = Variable([1 2; 3 4; 5 6]) name: nothing values: [1 2; 3 4; 5 6] creator: User-Defined(nothing)

julia> JITrench.flatten(x) name: nothing values: [1, 3, 5, 2, 4, 6] creator: JITrench.Flatten

JITrench.matmul — Method

matmul(x, W)

Matrix multiplication.

Example

julia> x = Variable([1 1; 0 1])
name: nothing 
values: [1 1; 0 1]
creator: User-Defined(nothing)

julia> W = Variable([1 0; 1 1])
name: nothing 
values: [1 0; 1 1]
creator: User-Defined(nothing)

julia> matmul(x, W)
name: nothing 
values: [2 1; 1 1]
creator: MatMul

JITrench.mean_squared_error — Method

mean_squared_error(y_true, y_pred)

Examples

julia> y_true = Variable([10, 20, 30])
name: nothing 
values: [10, 20, 30]
creator: User-Defined(nothing)

julia> y_pred = Variable([12, 20, 25])
name: nothing 
values: [12, 20, 25]
creator: User-Defined(nothing)

julia> mean_squared_error(y_true, y_pred)
name: nothing 
values: 9.666666666666666
creator: JITrench.MeanSquaredError

JITrench.plot_graph — Method

plot_graph(var::Variable; to_file="", title="")

Draw a computational graph with y as the end. This requires the installation of graphviz.

Arguments

to_file: Specifies the file name where the image will be saved.

If it is an empty string, it will return an image of type PNGContainer. See also: PNGContainer

title: Title of graph.

julia> x = Variable(10)
name: nothing 
values: 10
creator: User-Defined(nothing)

julia> y = x + 3
name: nothing 
values: 13
creator: JITrench.Add

julia> JITrench.plot_graph(y, to_file="graph.png") 

JITrench.plot_model — Method

plot_model(model::Model, inputs...,  to_file="model.png")

plot model by graphviz. Note that JITrench builds the graph at the time of calculation, so some kind of input is required.

JITrench.sum_to — Method

sum_to(x, shape)

Take the sum of each axis to form a shape.

Examples

julia> x = Variable(rand(2, 3))
name: nothing 
values: [0.2397911359535343 0.34270903251201146 0.699060178623987; 0.2345132451371843 0.21845435948625758 0.2924942369518322]
creator: User-Defined(nothing)

julia> JITrench.sum_to(x, (2, 1))
name: nothing 
values: [1.2815603470895327; 0.7454618415752741]
creator: JITrench.SumTo