API

AutoDiff

JITrench.AutoDiff.AbstractTensor — Type

AbstractTensor <: Variable

Abstract tensor variable. Super type of Tensor and CuTensor.

JITrench.AutoDiff.BinaryOperator — Type

BinaryOperator

DiffableFunction which takes two variables as input.

JITrench.AutoDiff.CPU — Type

CPU <: Device

Repersents CPU device. This is used for the default device.

JITrench.AutoDiff.GPU — Type

GPU(idx::Int64) <: Device

Repersents GPU device. idx is corresponding device index in CUDA.jl.

JITrench.AutoDiff.GradField — Type

GradField(inputs, output, generation)

Retain the information of the function for backpropagation. All DiffableFunction must have this field.

JITrench.AutoDiff.NotSameDeviceError — Type

NotSameDeviceError <: Exception

Exception thrown when the device of two tensors are not the same.

JITrench.AutoDiff.Scalar — Type

Scalar{T <: Real} <: Variable

Scalar variable.

JITrench.AutoDiff.Tensor — Type

Tensor{T <: AbstractArray} <: AbstractTensor

Tensor variable. This Variable is used for CPU.

JITrench.AutoDiff.UnaryOperator — Type

UnaryOperator

DiffableFunction which takes one variable as input.

JITrench.AutoDiff._get_gf — Method

_get_gf(f::DiffableFunction)

Get the GradField of the function.

JITrench.AutoDiff.backward! — Method

backward(y::Scalar; retain_grad = false, create_graph = false)

Clculate the gradient for y for all inputs. If retain_grad is true, all intermediate gradients are retained. If create_graph is true, the graph of the backward pass is constructed, allowing to compute higher order derivative products.

Arguments

y : Scalar which is the output of the function. Gradient can be immplicitly created only for Scalar.

retain_grad : If true, all intermediate gradients are retained.

create_graph : If true, the graph of the backward pass is constructed, allowing to compute higher order derivative products.

Returns

nothing: This function returns nothing. The gradient for each input is stored in grad field of each input.

Example

julia> x = Scalar(3)
Scalar{Int64}(3)

julia> y = Scalar(2.0)
Scalar{Float64}(2.0)

julia> z = x * 2y
Scalar{Float64}(12.0)

julia> backward!(z)

julia> x.grad
4.0

julia> y.grad
6.0

JITrench.AutoDiff.backward — Method

backward(f::DiffableFunction, gy)

Definition of backward calculation of F. This function must be implemented for each DiffableFunction.

Arguments

F::DiffableFunction : Instance of DiffableFunction which is called.

gy : propageted gradient from output.

Returns

gxs... : propageted gradient to inputs.

Example

julia>julia> JITrench.AutoDiff.backward(f, 1)
(1, 1)

JITrench.AutoDiff.call! — Function

call!(F::Type{<:DiffableFunction}, xs...::Variable)

Call JITrench function with xs... as inputs.

JITrench.AutoDiff.call! — Method

call!(F::Type{<:DiffableFunction}, xs...::Variable)

Call JITrench function with xs... as inputs.

Arguments

F::Type{<:DiffableFunction} : Type of DiffableFunction which is called.

xs...::Variable : Variable which is passed to F.

Returns

y : values of output.

Example

julia> x = Scalar(3)
Scalar{Int64}(3)

julia> y = Scalar(4)
Scalar{Int64}(4)

julia> JITrench.AutoDiff.call!(JITrench.Add, x, y)
Scalar{Int64}(7)

JITrench.AutoDiff.call! — Method

call!(F::Type{<:DiffableFunction}, additional_field::AdditionalField, xs...::Variable)

Call JITrench function with xs... as inputs, and additional_field as optional argument.

Arguments

F::Type{<:DiffableFunction} : Type of DiffableFunction which is called.

additional_field::AdditionalField : AdditionalField which is passed to F.

xs...::Variable : Variable which is passed to F.

Returns

y : values of output.

Example

julia> x = Scalar(3)
Scalar{Int64}(3)

julia> y = JITrench.AutoDiff.call!(JITrench.Pow, JITrench.PowField(3), x)
Scalar{Int64}(27)

julia> y = JITrench.AutoDiff.call!(JITrench.Pow, JITrench.PowField(2), x)
Scalar{Int64}(9)

JITrench.AutoDiff.check_same_device — Method

check_same_device(device1::Device, device2::Device)

Check if device1 and device2 are the same device. If they are the same device, return nothing. Otherwise, throw NotSameDeviceError.

Arguments

device1: Device to be compared.

device2: Device to be compared.

Example

julia> device1 = JITrench.CPU()
JITrench.AutoDiff.CPU()

julia> device2 = JITrench.GPU(0)
JITrench.AutoDiff.GPU(0)

julia> JITrench.check_same_device(device1, device2)
ERROR: All arguments must be in the same device. Arguments are on different GPUs.

JITrench.AutoDiff.forward — Method

forward(F::Type{DiffableFunction}, args...)

Definition of forward calculation of F. This function must be implemented for each DiffableFunction. This function is called by call! function with values of inputs.

Arguments

F::Type{DiffableFunction} : Type of DiffableFunction which is called.

args... : values of inputs.

Returns

y : values of output.

Example

julia>julia> JITrench.AutoDiff.forward(JITrench.Add, 1, 2)
3

JITrench.AutoDiff.get_values — Method

get_values(x)

Get values from x. If x is Variable, return x.values. Otherwise, return x.

JITrench.AutoDiff.make_func — Function

make_func(::Type{F}, y, inputs, gen)

Create DiffableFunction instance from y, inputs and gen.

JITrench.AutoDiff.make_func — Method

make_func(::Type{F}, y, inputs, gen)

Create DiffableFunction instance from y, inputs and gen.

Arguments

F::Type{F} : Type of DiffableFunction which is created.

y : values of output.

inputs : inputs of F.

gen : generation of output.

JITrench.AutoDiff.make_func — Method

make_func(::Type{F}, additional_field, y, inputs, gen)

Create DiffableFunction instance from additional_field, y, inputs and gen.

Arguments

F::Type{F} : Type of DiffableFunction which is created.

additional_field : additional_field for F.

y : values of output.

inputs : inputs of F.

gen : generation of output.

JITrench.AutoDiff.make_func — Method

make_func(::Type{F}, y::CuArray, inputs, gen, device_idx)

Create DiffableFunction instance from y, inputs , gen and device_idx.

Arguments

F::Type{F} : Type of DiffableFunction which is created.

y : values of output which is CuArray.

inputs : inputs of F.

gen : generation of output.

JITrench.AutoDiff.out_to_tensor — Function

out_to_tensor(y, generation; creator=nothing, grad=nothing, req_broadcast=false)

Convert output of forward function to appropriate Variable. (e.g. Scalar, Tensor, CuTensor)

Arguments

y : values of output.

generation : generation of output.

creator : creator of output.

grad : grad of output.

req_broadcast : req_broadcast of output.

Returns

out : Variable which is converted from y.

JITrench.AutoDiff.@diffable — Macro

@diffable

Check definition of DiffableFunction has grad_field field and it's type is GradField.

JITrench.AutoDiff.@nograd — Macro

@nograd

Disable gradient calculation in the scope of the macro.

Example

julia> x = Tensor([1, 2, 3])
3-element Tensor{Vector{Int64}}:
 1
 2
 3

julia> y = 2x
3-element Tensor{Vector{Int64}}:
 2
 4
 6

julia> JITrench.AutoDiff.@nograd y = 2x
3-element Vector{Int64}:
 2
 4
 6

julia> JITrench.AutoDiff.@nograd begin
           y = 2x
       end
3-element Vector{Int64}:
 2
 4
 6

Utilities

JITrench.plot_graph — Function

plot_graph(var :: Variable; to_file="", title="")

Draw a computational graph with y as the end. This requires the installation of graphviz.

Arguments

to_file: Specifies the file name where the image will be saved.

If it is an empty string, it will return an image of type PNGContainer. See also: PNGContainer

title: Title of graph.

julia> x = Variable(10)
name: nothing 
values: 10
creator: User-Defined(nothing)

julia> y = x + 3
name: nothing 
values: 13
creator: JITrench.Add

julia> JITrench.plot_graph(y, to_file="graph.png") 

JITrench.PNGContainer — Type

PNGContainer

A structure for handling images. You can display an image by using it as a return value of a cell or by explicitly display( :: PNGContainer) in Jupyter.