JITrench

JITrench.PNGContainer

JITrench.plot_graph

JITrench.PNGContainer — Type

PNGContainer

A structure for handling images. You can display an image by using it as a return value of a cell or by explicitly display( :: PNGContainer) in Jupyter.

JITrench.plot_graph — Method

plot_graph(var :: Variable; to_file="", title="")

Draw a computational graph with y as the end. This requires the installation of graphviz.

Arguments

to_file: Specifies the file name where the image will be saved.

If it is an empty string, it will return an image of type PNGContainer. See also: PNGContainer

title: Title of graph.

julia> x = Variable(10)
name: nothing 
values: 10
creator: User-Defined(nothing)

julia> y = x + 3
name: nothing 
values: 13
creator: JITrench.Add

julia> JITrench.plot_graph(y, to_file="graph.png")